requirements.txt
aniso8601==7.0.0
arrow==0.14.5
astroid==2.2.5
certifi==2019.6.16
chardet==3.0.4
Click==7.0
coloredlogs==10.0
Flask==1.0.2
Flask-Cors==3.0.8
Flask-RESTful==0.3.7
Flask-Script==2.0.6
humanfriendly==4.18
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
kafka-python==1.4.6
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==2.3.1
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.2
requests==2.22.0
six==1.12.0
typed-ast==1.4.0
urllib3==1.25.3
Werkzeug==0.15.5
wrapt==1.11.2
